add mockerRedisClient to test folder and migrate normalize functionality to separate cacheOperation
